fixed sidebar logical location
fixed issue that sidebar cant open when there are no messages in the channel
