+++ packages/react/src/components/ChatBody/ChatBody.js
@@ -3,7 +3,18 @@ import React, { useCallback, useContext, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { css } from '@emotion/react';
 import RCContext from '../../context/RCInstance';
-import { useMessageStore, useUserStore, useChannelStore } from '../../store';
+import {
+  useMessageStore,
+  useUserStore,
+  useChannelStore,
+  usePinnedMessageStore,
+  useStarredMessageStore,
+  useSearchMessageStore,
+  useFileStore,
+  useMentionsStore,
+  useThreadsMessageStore,
+  useMemberStore,
+} from '../../store';
 import MessageList from '../MessageList';
 import TotpModal from '../TotpModal/TwoFactorTotpModal';
 import { Box } from '../Box';
@@ -14,6 +25,14 @@ import ModalBlock from '../uiKit/blocks/ModalBlock';
 import useComponentOverrides from '../../theme/useComponentOverrides';
 import RecentMessageButton from './RecentMessageButton';
 import useFetchChatData from '../../hooks/useFetchChatData';
+import RoomMembers from '../RoomMembers/RoomMember';
+import UserMentions from '../UserMentions/UserMentions';
+import AllThreads from '../AllThreads/AllThreads';
+import PinnedMessages from '../PinnedMessages/PinnedMessages';
+import StarredMessages from '../StarredMessages/StarredMessages';
+import SearchMessage from '../SearchMessage/SearchMessage';
+import Roominfo from '../RoomInformation/RoomInformation';
+import { Files } from '../Files';
 
 const ChatBody = ({
   height,
@@ -62,6 +81,18 @@ const ChatBody = ({
   const upsertMessage = useMessageStore((state) => state.upsertMessage);
   const removeMessage = useMessageStore((state) => state.removeMessage);
   const isChannelPrivate = useChannelStore((state) => state.isChannelPrivate);
+
+  const showMentions = useMentionsStore((state) => state.showMentions);
+  const showAllFiles = useFileStore((state) => state.showAllFiles);
+  const showAllThreads = useThreadsMessageStore(
+    (state) => state.showAllThreads
+  );
+  const showPinned = usePinnedMessageStore((state) => state.showPinned);
+  const showStarred = useStarredMessageStore((state) => state.showStarred);
+  const showSearch = useSearchMessageStore((state) => state.showSearch);
+  const showChannelinfo = useChannelStore((state) => state.showChannelinfo);
+  const showMembers = useMemberStore((state) => state.showMembers);
+  const members = useMemberStore((state) => state.members);
 
   const [isThreadOpen, threadMainMessage] = useMessageStore((state) => [
     state.isThreadOpen,
@@ -268,6 +299,15 @@ const ChatBody = ({
           onClick={handlePopupClick}
         />
       )}
+
+      {showMembers && <RoomMembers members={members} />}
+      {showSearch && <SearchMessage />}
+      {showChannelinfo && <Roominfo />}
+      {showAllThreads && <AllThreads />}
+      {showAllFiles && <Files />}
+      {showMentions && <UserMentions />}
+      {showPinned && <PinnedMessages />}
+      {showStarred && <StarredMessages />}
     </>
   );
 };
