+++ packages/react/src/components/MessageList/MessageList.js
@@ -1,47 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { isSameDay } from 'date-fns';
-import {
-  useMessageStore,
-  useMemberStore,
-  useSearchMessageStore,
-  useChannelStore,
-  useUserStore,
-  useFileStore,
-  useMentionsStore,
-  useThreadsMessageStore,
-  usePinnedMessageStore,
-  useStarredMessageStore,
-} from '../../store';
-import RoomMembers from '../RoomMembers/RoomMember';
+import { useMessageStore, useUserStore } from '../../store';
 import MessageReportWindow from '../ReportMessage/MessageReportWindow';
 import isMessageSequential from '../../lib/isMessageSequential';
-import UserMentions from '../UserMentions/UserMentions';
-import SearchMessage from '../SearchMessage/SearchMessage';
-import Roominfo from '../RoomInformation/RoomInformation';
-import AllThreads from '../AllThreads/AllThreads';
-import { Files } from '../Files';
 import { Message } from '../Message';
-import PinnedMessages from '../PinnedMessages/PinnedMessages';
-import StarredMessages from '../StarredMessages/StarredMessages';
+
 import { Icon } from '../Icon';
 
 const MessageList = ({ messages }) => {
-  const showSearch = useSearchMessageStore((state) => state.showSearch);
-  const showChannelinfo = useChannelStore((state) => state.showChannelinfo);
-  const showMembers = useMemberStore((state) => state.showMembers);
-  const members = useMemberStore((state) => state.members);
   const showReportMessage = useMessageStore((state) => state.showReportMessage);
   const messageToReport = useMessageStore((state) => state.messageToReport);
   const showAvatar = useUserStore((state) => state.showAvatar);
   const headerTitle = useMessageStore((state) => state.headerTitle);
-  const showMentions = useMentionsStore((state) => state.showMentions);
-  const showAllFiles = useFileStore((state) => state.showAllFiles);
-  const showAllThreads = useThreadsMessageStore(
-    (state) => state.showAllThreads
-  );
-  const showPinned = usePinnedMessageStore((state) => state.showPinned);
-  const showStarred = useStarredMessageStore((state) => state.showStarred);
 
   const isMessageNewDay = (current, previous) =>
     !previous || !isSameDay(new Date(current.ts), new Date(previous.ts));
@@ -85,15 +56,7 @@ const MessageList = ({ messages }) => {
           );
         })}
 
-      {showMembers && <RoomMembers members={members} />}
       {showReportMessage && <MessageReportWindow messageId={messageToReport} />}
-      {showSearch && <SearchMessage />}
-      {showChannelinfo && <Roominfo />}
-      {showAllThreads && <AllThreads />}
-      {showAllFiles && <Files />}
-      {showMentions && <UserMentions />}
-      {showPinned && <PinnedMessages />}
-      {showStarred && <StarredMessages />}
     </>
   );
 };
